requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic
python-dotenv
redis
geoalchemy2
geopy
pydantic-settings
